refactor: remove index.js for general components
