+++ src/app/(main)/(components)/card-list/index.jsx
@@ -1,4 +1,7 @@
-import { Button, Card, Icon } from "@/components";
+import Button from "@/components/button";
+import Card from "@/components/card";
+import Icon from "@/components/icon";
+
 import styles from "./styles.module.scss";
 import Link from "next/link";
 import { useUserAgent } from "@/tools/user-agent";
