+++ src/app/(main)/page.js
@@ -1,4 +1,4 @@
-import { Header } from "@/components";
+import Header from "@/components/header";
 import Banner from "./(components)/banner";
 import CardList from "./(components)/card-list";
 import Indexes from "./(components)/indexes";
