--- src/components/index.js
@@ -1,8 +1,0 @@
-export { default as Button } from "./button";
-export { default as Card } from "./card";
-export { default as Checkbox } from "./checkbox";
-export { default as Footer } from "./footer";
-export { default as Header } from "./header";
-export { default as Icon } from "./icon";
-export { default as Logo } from "./logo";
-export { default as Textfield } from "./textfield";
